chore: fixed bug 'variable size keys index out of bounds error'
