+++ cmd/radFS/arttest/main.go
@@ -7,9 +7,10 @@ import (
 func main() {
 	var t art.Tree
 
-	t.Insert([]byte("cat"), "v1")
-	t.Insert([]byte("car"), "v2")
-	t.Insert([]byte("cap"), "v3")
+	t.Insert([]byte("cats"), "v1")
+
+	t.Insert([]byte("cat"), "v3")
+	t.Insert([]byte("carpet"), "v3")
 
 	art.PrintTree(t.Root(), 0)
 }
