+++ internal/art/print_tree.go
@@ -19,19 +19,19 @@ func PrintTree(n *Node, level int) {
 
 	in := n.innerNode
 
-	prefixLen := in.meta.prefixlen
-	if prefixLen < 0 || prefixLen > len(in.meta.prefix) {
-		prefixLen = 0
+	prefixlen := in.meta.prefixlen
+	if prefixlen < 0 || prefixlen > len(in.meta.prefix) {
+		prefixlen = 0
 	}
 
-	prefix := string(in.meta.prefix[:prefixLen])
+	prefix := string(in.meta.prefix[:prefixlen])
 
-	fmt.Println(indent+"Node(prefix=\""+prefix+"\", prefixLen=", prefixLen, ")")
+	fmt.Println(indent+"Node(prefix=\""+prefix+"\", prefixLen=", prefixlen, ")")
 
 	// Print children
 	for i := 0; i < len(in.keys); i++ {
 		if in.children[i] != nil {
-			fmt.Println(indent+" Edge('", string(in.keys[i]), "'):")
+			fmt.Printf("%s Edge('%c' | %d):\t", indent, in.keys[i], in.keys[i])
 			PrintTree(in.children[i], level+1)
 		}
 	}
